requirements.txt
numpy
pandas
scikit-learn
python-dotenv
fastapi
uvicorn
pymongo
